requirements.txt
pandas
numpy
scipy
python-dotenv
requests
tqdm
